tests/wpt/mozilla/tests/mozilla/create_element.html

<html>
<head>
  <title></title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
</head>
<body>
  <script>
  test(function() {
    var elem = document.createElement("foo");
    assert_equals(elem.tagName, "FOO");
    var elem = document.createElement("p");
    assert_true(elem instanceof HTMLParagraphElement, "Should be HTMLParagraphElement");
    var elem = document.createElement("sPAn");
    assert_true(elem instanceof HTMLSpanElement, "Should be HTMLSpanElement");
    var text = document.createTextNode("hello");
    assert_true(text instanceof Text, "Should be Text");
  });
  </script>
</body>
</html>
